[FIX] mail: let auto_subscribe use up-to-date values
`_message_auto_subscribe` relies on `_get_subscription_data` to know who
is subscribed to what. This method uses a raw SQL query but doesn't flush
the relevant records beforehand.

This is not an issue in most normal cases, but if followers are edited manually
for example in order to merge existing records. It may cause issues when
assigning a user to the record after merging, as the this will attempt to add
them to the list of followers automatically, when they already were on it.

task-3713472
original issue in opw-3634410

closes odoo/odoo#155144

X-original-commit: fa1fdc2a188755cc5df3153c630cb2e1c6453f56
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>